fix: backend type/schema DRY — PeriodType, Numeric hints, dict access (#336)
