+++ backend/app/models/price.py
@@ -13,10 +13,10 @@ class PriceHistory(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     asset_id: Mapped[int] = mapped_column(ForeignKey("assets.id", ondelete="CASCADE"), index=True)
     date: Mapped[date] = mapped_column(Date, index=True)
-    open: Mapped[float] = mapped_column(Numeric(12, 4))
-    high: Mapped[float] = mapped_column(Numeric(12, 4))
-    low: Mapped[float] = mapped_column(Numeric(12, 4))
-    close: Mapped[float] = mapped_column(Numeric(12, 4))
+    open: Mapped[float] = mapped_column(Numeric(12, 4, asdecimal=False))
+    high: Mapped[float] = mapped_column(Numeric(12, 4, asdecimal=False))
+    low: Mapped[float] = mapped_column(Numeric(12, 4, asdecimal=False))
+    close: Mapped[float] = mapped_column(Numeric(12, 4, asdecimal=False))
     volume: Mapped[int] = mapped_column(BigInteger)
 
     asset: Mapped["Asset"] = relationship(back_populates="prices")
